+++ plugins/catalog-backend/src/service/CatalogBuilder.ts
@@ -56,7 +56,7 @@ import {
   yamlPlaceholderResolver,
 } from '../modules/core/PlaceholderProcessor';
 import { defaultEntityDataParser } from '../modules/util/parse';
-import { LocationAnalyzer } from '../ingestion/types';
+import { LocationAnalyzer, ScmLocationAnalyzer } from '../ingestion/types';
 import { CatalogProcessingEngine } from '../processing';
 import { DefaultProcessingDatabase } from '../database/DefaultProcessingDatabase';
 import { applyDatabaseMigrations } from '../database/migrations';
@@ -115,6 +115,10 @@ export type CatalogEnvironment = {
  *   after the processors' pre-processing steps. All policies are given the
  *   chance to inspect the entity, and all of them have to pass in order for
  *   the entity to be considered valid from an overall point of view.
+ * - Location analyzers can be added. These are responsible for analyzing
+ *   repositories when onboarding them into the catalog, by finding
+ *   catalog-info.yaml files and other artifacts that can help automatically
+ *   register or create catalog data on the user's behalf.
  * - Placeholder resolvers can be replaced or added. These run on the raw
  *   structured data between the parsing and pre-processing steps, to replace
  *   dollar-prefixed entries with their actual values (like $file).
@@ -135,6 +139,7 @@ export class CatalogBuilder {
   private fieldFormatValidators: Partial<Validators>;
   private entityProviders: EntityProvider[];
   private processors: CatalogProcessor[];
+  private locationAnalyzers: ScmLocationAnalyzer[];
   private processorsReplace: boolean;
   private parser: CatalogProcessorParser | undefined;
   private onProcessingError?: (event: {
@@ -165,6 +170,7 @@ export class CatalogBuilder {
     this.fieldFormatValidators = {};
     this.entityProviders = [];
     this.processors = [];
+    this.locationAnalyzers = [];
     this.processorsReplace = false;
     this.parser = undefined;
     this.permissionRules = Object.values(catalogPermissionRules);
@@ -336,6 +342,21 @@ export class CatalogBuilder {
   }
 
   /**
+   * Adds Location Analyzers. These are responsible for analyzing
+   * repositories when onboarding them into the catalog, by finding
+   * catalog-info.yaml files and other artifacts that can help automatically
+   * register or create catalog data on the user's behalf.
+   *
+   * @param locationAnalyzers - One or more location analyzers
+   */
+  addLocationAnalyzers(
+    ...analyzers: Array<ScmLocationAnalyzer | Array<ScmLocationAnalyzer>>
+  ): CatalogBuilder {
+    this.locationAnalyzers.push(...analyzers.flat());
+    return this;
+  }
+
+  /**
    * Sets up the catalog to use a custom parser for entity data.
    *
    * This is the function that gets called immediately after some raw entity
@@ -478,7 +499,8 @@ export class CatalogBuilder {
     );
 
     const locationAnalyzer =
-      this.locationAnalyzer ?? new RepoLocationAnalyzer(logger, integrations);
+      this.locationAnalyzer ??
+      new RepoLocationAnalyzer(logger, integrations, this.locationAnalyzers);
     const locationService = new AuthorizedLocationService(
       new DefaultLocationService(locationStore, orchestrator, {
         allowedLocationTypes: this.allowedLocationType,