+++ packages/integration-react/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/integration-react
 
+## 1.1.5-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/config@1.0.3-next.2
+  - @backstage/core-components@0.11.2-next.2
+  - @backstage/core-plugin-api@1.0.7-next.2
+  - @backstage/integration@1.3.2-next.2
+  - @backstage/theme@0.2.16
+
 ## 1.1.5-next.1
 
 ### Patch Changes