+++ packages/test-utils/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/test-utils
 
+## 1.2.1-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-permission-common@0.7.0-next.2
+  - @backstage/plugin-permission-react@0.4.6-next.2
+  - @backstage/config@1.0.3-next.2
+  - @backstage/core-app-api@1.1.1-next.2
+  - @backstage/core-plugin-api@1.0.7-next.2
+  - @backstage/theme@0.2.16
+  - @backstage/types@1.0.0
+
 ## 1.2.1-next.1
 
 ### Patch Changes