+++ plugins/apache-airflow/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/plugin-apache-airflow
 
+## 0.2.3-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.11.2-next.2
+  - @backstage/core-plugin-api@1.0.7-next.2
+
 ## 0.2.3-next.1
 
 ### Patch Changes