+++ plugins/example-todo-list/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @internal/plugin-todo-list
 
+## 1.0.6-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.11.2-next.2
+  - @backstage/core-plugin-api@1.0.7-next.2
+  - @backstage/theme@0.2.16
+
 ## 1.0.6-next.1
 
 ### Patch Changes