+++ plugins/github-actions/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-github-actions
 
+## 0.5.10-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog-react@1.2.0-next.2
+  - @backstage/catalog-model@1.1.2-next.2
+  - @backstage/core-components@0.11.2-next.2
+  - @backstage/core-plugin-api@1.0.7-next.2
+  - @backstage/integration@1.3.2-next.2
+  - @backstage/theme@0.2.16
+
 ## 0.5.10-next.1
 
 ### Patch Changes