+++ plugins/techdocs-backend/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @backstage/plugin-techdocs-backend
 
+## 1.4.0-next.2
+
+### Minor Changes
+
+- 7ced1b4076: Add optional `catalogClient` argument to `createRoute` parameters
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog-common@1.0.7-next.2
+  - @backstage/backend-common@0.15.2-next.2
+  - @backstage/plugin-permission-common@0.7.0-next.2
+  - @backstage/plugin-techdocs-node@1.4.1-next.2
+  - @backstage/plugin-search-common@1.1.0-next.2
+  - @backstage/catalog-client@1.1.1-next.2
+  - @backstage/catalog-model@1.1.2-next.2
+  - @backstage/config@1.0.3-next.2
+  - @backstage/errors@1.1.2-next.2
+  - @backstage/integration@1.3.2-next.2
+
 ## 1.3.1-next.1
 
 ### Patch Changes